fix(app): unify authorization-server responses and log callback failures
Both OAuth handlers had grown their own copy of the query limits, the
duplicate-parameter check, the trusted-redirect builder, and the response
helpers, and the copies had already drifted: `authorize`'s direct errors
omitted the `referrer-policy: no-referrer` that `callback`'s set. Sharing
them through `authorization_server::{query, response}` closes that omission
and leaves one place for the next handler to reuse rather than a third copy
to keep in step.

The missing header survived because `authorize` only asserted headers on its
redirect path, where they were present; its direct-error tests checked the
status and nothing else. Both paths now assert the whole set, so the next
divergence fails a test instead of waiting for a reader to notice.

The callback also discarded every typed error it had. An operator debugging
a broken IdP saw `error=server_error` in the browser and nothing at all
server-side, even though the provider-client and state-store errors name the
stage that failed and carry no provider-supplied data. All four
`server_error` paths now record theirs.

Also narrows `OIDC_CALLBACK_PATH` to `pub(in crate::auth)`, records why the
callback's `Url::parse` fallback is unreachable, and documents the reserved
query parameter a registered redirect URI may carry, which client
registration is where to reject.
